+++ src/background/keyboard.js
@@ -2,6 +2,7 @@
 
 import Items      from '../modules/items.js';
 import Keyboard   from '../modules/keyboard.js';
+import Logger     from '../modules/logger.js';
 import Settings   from '../modules/settings.js';
 import { KeyboardShortcuts } from '../modules/constants.js';
 
@@ -13,32 +14,40 @@ import { KeyboardShortcuts } from '../modules/constants.js';
 // https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/commands/update
 if(browser.commands.update) {
   Settings.init().then( function() {
-    let settings = Settings.get();
+    const settings = Settings.get();
 
     Keyboard.registerShortcut(KeyboardShortcuts.toggle,    settings.keyboardToggle);
     Keyboard.registerShortcut(KeyboardShortcuts.openPopup, settings.keyboardOpenPopup);
+    Keyboard.registerShortcut(KeyboardShortcuts.openFirstItem,  settings.keyboardOpenFirstItem);
+    Keyboard.registerShortcut(KeyboardShortcuts.openRandomItem, settings.keyboardOpenRandomItem);
   });
 }
 
+browser.commands.onCommand.addListener( command => {
+  switch(command) {
+    case KeyboardShortcuts.toggle:
+      Logger.log('(keyboard) KeyboardShortcuts.toggle');
+      browser.tabs.query({ active: true, currentWindow: true }).then( ([currentTab]) => {
+        // FIXME: duplication with PageAction.toggle())
+        browser.storage.local.get('items').then( ({ items }) => {
+          const matchingItem = Items.find( items, { url: currentTab.url });
 
-browser.commands.onCommand.addListener( (command) => {
-  if( command === KeyboardShortcuts.toggle ) {
-    browser.tabs.query({ active: true, currentWindow: true }).then( ([currentTab]) => {
-      //
-      // PageAction.toggle(currentTab);
-      //
-
-      // FIXME: duplication with PageACtion.toggle())
-      browser.storage.local.get('items').then( ({ items }) => {
-        const matchingItem = Items.find( items, { url: currentTab.url });
-
-        if( matchingItem ) {
-          Items.markAsRead( matchingItem.id );
-        } else {
-          const addItemOptions = { closeTabId: currentTab.id };
-          Items.addItem( currentTab.url, currentTab.title, addItemOptions );
-        }
+          if(matchingItem) {
+            Items.markAsRead(matchingItem.id);
+          } else {
+            const addItemOptions = { closeTabId: currentTab.id };
+            Items.addItem(currentTab.url, currentTab.title, addItemOptions);
+          }
+        });
       });
-    });
+      break;
+    case KeyboardShortcuts.openFirstItem:
+      Logger.log('(keyboard) KeyboardShortcuts.openFirstItem');
+      Items.openFirst();
+      break;
+    case KeyboardShortcuts.openRandomItem:
+      Logger.log('(keyboard) KeyboardShortcuts.openRandomItem');
+      Items.openRandom();
+      break;
   }
 });